Melhoria nas Condições e Atualização de algumas condição da aplicação.

+++ src/services/DeleteUserService.js
@@ -2,8 +2,13 @@ const UserRepository = require('../repositories/UserRepository');
 
 class DeleteUserService {
 
-  execute = async ({ login }) => {
+  execute = async ({ login, role, user_id }) => {
     const userRepository = new UserRepository();
+    const userFound = await userRepository.findById({ user_id });
+
+    if (userFound.role !== 'admin' && role !== 'voter') {
+      throw new Error('Only Administrators can perform this action');
+    }
 
     const deleteddUser = await userRepository.delete({ login });
